python/meli_be/requirements.txt
asgiref==3.6.0
astroid==2.15.8
attrs==22.2.0
black==23.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
dill==0.3.7
distlib==0.3.6
Django==4.1.7
django-filter==22.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docutils==0.20.1
filelock==3.10.0
idna==3.6
iniconfig==2.0.0
isort==5.12.0
lazy-object-proxy==1.9.0
Markdown==3.4.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.0.0
PyJWT==2.6.0
pylint==2.17.6
pytest==7.2.2
pytest-django==4.5.2
pytz==2022.7.1
requests==2.31.0
rjsmin==1.2.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
sqlparse==0.4.3
text-unidecode==1.3
tomlkit==0.12.1
urllib3==2.2.1
virtualenv==20.21.0
wcwidth==0.2.5
wrapt==1.15.0
